# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Linc Corporate Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TrainingProgram {
  id: string;
  title: string;
  description: string;
  category: string;
  duration: string;
  level?: string;
  outcomes: string[];
  benefits: string[];
  /** @nullable */
  imageUrl?: string | null;
  featured?: boolean;
  createdAt?: string;
}

export interface TrainingProgramInput {
  title: string;
  description: string;
  category: string;
  duration: string;
  level?: string;
  outcomes: string[];
  benefits: string[];
  imageUrl?: string;
  featured?: boolean;
}

export interface Trainer {
  id: string;
  name: string;
  title: string;
  category: string;
  bio: string;
  expertise: string[];
  experience?: string;
  availability: string;
  rating?: number;
  sessionsCompleted?: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkedin?: string | null;
  featured?: boolean;
  createdAt?: string;
}

export interface TrainerInput {
  name: string;
  title: string;
  category: string;
  bio: string;
  expertise: string[];
  experience?: string;
  availability: string;
  rating?: number;
  sessionsCompleted?: number;
  imageUrl?: string;
  linkedin?: string;
  featured?: boolean;
}

export interface Inquiry {
  id: string;
  name: string;
  email: string;
  company: string;
  type: string;
  message: string;
  /** @nullable */
  phone?: string | null;
  status: string;
  createdAt: string;
}

export interface InquiryInput {
  name: string;
  email: string;
  company: string;
  type: string;
  message: string;
  phone?: string;
}

export type PortalStatsTrainerCategoriesItem = {
  category: string;
  count: number;
};

export interface PortalStats {
  totalTrainers: number;
  totalPrograms: number;
  totalInquiries: number;
  featuredTrainers: number;
  featuredPrograms: number;
  trainerCategories: PortalStatsTrainerCategoriesItem[];
}

export type ListTrainersParams = {
category?: string;
};