# lib/api-zod/src/generated/types/inquiry.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Linc Corporate Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Inquiry {
  id: string;
  name: string;
  email: string;
  company: string;
  type: string;
  message: string;
  /** @nullable */
  phone?: string | null;
  status: string;
  createdAt: string;
}