# lib/api-zod/src/generated/types/portalStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Linc Corporate Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { PortalStatsTrainerCategoriesItem } from './portalStatsTrainerCategoriesItem';

export interface PortalStats {
  totalTrainers: number;
  totalPrograms: number;
  totalInquiries: number;
  featuredTrainers: number;
  featuredPrograms: number;
  trainerCategories: PortalStatsTrainerCategoriesItem[];
}