# lib/api-zod/src/generated/types/trainer.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * AI Linc Corporate Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Trainer {
  id: string;
  name: string;
  title: string;
  category: string;
  bio: string;
  expertise: string[];
  experience?: string;
  availability: string;
  rating?: number;
  sessionsCompleted?: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkedin?: string | null;
  featured?: boolean;
  createdAt?: string;
}